+++ sentry/src/main/java/io/sentry/NoOpTransactionProfiler.java
@@ -1,5 +1,6 @@
 package io.sentry;
 
+import java.util.List;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
@@ -18,7 +19,8 @@ public final class NoOpTransactionProfiler implements ITransactionProfiler {
 
   @Override
   public @Nullable ProfilingTraceData onTransactionFinish(
-      @NotNull ITransaction transaction, @Nullable PerformanceCollectionData memoryCollectionData) {
+      @NotNull ITransaction transaction,
+      @Nullable List<PerformanceCollectionData> performanceCollectionData) {
     return null;
   }
 }