+++ sentry/src/test/java/io/sentry/SentryOptionsTest.kt
@@ -388,7 +388,21 @@ class SentryOptionsTest {
         assertEquals("${File.separator}test${File.separator}${hash}${File.separator}profiling_traces", options.profilingTracesDirPath)
     }
 
+    @Test
     fun `when options are initialized, idleTimeout is 3000`() {
         assertEquals(3000L, SentryOptions().idleTimeout)
+    }
+
+    @Test
+    fun `when options are initialized, TransactionPerformanceCollector is a NoOp`() {
+        assertEquals(SentryOptions().transactionPerformanceCollector, NoOpTransactionPerformanceCollector.getInstance())
+    }
+
+    @Test
+    fun `when setTransactionPerformanceCollector is called, overrides default`() {
+        val performanceCollector = mock<TransactionPerformanceCollector>()
+        val options = SentryOptions()
+        options.transactionPerformanceCollector = performanceCollector
+        assertEquals(performanceCollector, options.transactionPerformanceCollector)
     }
 }